now using separate fixture files, slightly modified test setup; added a travis conf
