+++ test/fixtures/__init__.py
@@ -1,0 +1,3 @@
+from fixtures.my_func import my_func
+from fixtures.my_func_bad import my_func_bad
+
