+++ test/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+python_files=test/test_*.py
+
