+++ test/requirements.txt
@@ -1,0 +1,4 @@
+pytest==4.4.1
+
+# install citeme from local files
+.
